MAJOR changes: prototyping alternate logic of FeynWann e-ph wherein all transforms are done with the phonon supercell size.
This massively simplifies the ePhLoop logic, redefines the meaning of an "offset" for ePhLoop, and exposes much more parallelism.

Tested to work with phononElectronLinewidth, and requires testing of other executables.

Current version of phononElectronLinewidth seems to slow down due to increased calls to transform1: needs optimization of that logic in DistributedMatrix