Changed meaning of ePhDelta to be a standard Gaussian sigma so that delta ~ exp(-0.5*(DeltaE/ePhDelta)^2), whereas previously that was the exponential in sqrt(delta).
Effectively, this means that for the same simulation, the new code requires an ePhDelta which is sqrt(2) smaller than the old code.

Note that pumpTau and probe tau meanings have not changed: they refer to standard Gaussian sigmas of the amplitude (not the intensity). This has been clarified in a comment.